move the functions start, end, printBoard and info to src/core/game.js and the function move to src/core/snakes.js
